ODC-7632: Remove redundant navigation hrefs
Also run `yarn i18n`
